add a exit event listener to popup
Added an event listener to the newBookContainer variable that closes the popup when the backdrop is clicked. Implemented event.stopPropagation() on the form element to prevent the popup from closing when the user interacts with the form itself, ensuring only clicks on the backdrop trigger the close action. did this by using a if statement and said stopprognation of the event.target.id does not equal id "newBook"